Revert "file(GENERATE): Evaluate early to allow generating source files"
This reverts commit 28fedb192175668668e02e0bfd47fde337b83899.

Aside from the requirement to appear before TraceDependencies, there is
a separate requirement to appear after it, because it may change the
resulting content.

The solution is probably to try to mark GENERATED output files as such,
but I'd have to investigate to see if TraceDependencies can cause different
genex evaluation before and after, and whether that causes future
limitations.