cmDefinitions: only pull variables if they aren't local
Prior to the commit which changed behavior here, PARENT_SCOPE triggered
a ->GetInternal() call which would internally pull down the parent's
value into the local scope. When converted over to an explicit ->Pull()
call, the logic in ->GetInternal() which ignored the parent if it is
already in scope was missed and instead unconditionally pulled down.

Reported-By: Ben Cooksley <bcooksley@kde.org>