Revert "cmDefinitions: Don't store parent lookups"
This reverts commit 5abfde6cb8a1ae0b2825797eab6c2e9842eb7c49.

The behaviors associated with implicit pulldown on variable lookup
seriously conflict with the optimizations made in these commits.
Basically, since values were copied upon variable lookup, not just on
PARENT_SCOPE, coupled with PARENT_SCOPE's behavior based on whether the
variable is in the current scope or not causes serious problems with not
storing a value for every variable at every scope.

Reported-By: Ben Cooksley <bcooksley@kde.org>