Export: Diagnose issues in INCLUDES DESTINATION while generating (#15258).
The _IMPORT_PREFIX variable is generated in export files only if the
installation destination is relative.  If packagers set the path to
an absolute destination, the generated file must not have any
references to the _IMPORT_PREFIX as it is undefined.

A use of _IMPORT_PREFIX is generated for content passed to
install(TARGETS) in the INCLUDES DESTINATION component, if the
content is a relative path.  Issue a message to notify the user that
that the generated file is unusable, and introduce policy CMP0057 to
control the severity of the message.