Export: Issue error if generating unusable export file (#15258).
Diagnose presence of $<INSTALL_PREFIX> in exported target properties,
as can appear in user code, or generated for relative paths inside
a $<INSTALL_INTERFACE> generator expression.

The case of use of install(TARGETS) with a relative path in
INCLUDES DESTINATION would also be caught by this diagnostic, but
that is handled separately because in that case the $<INSTALL_PREFIX>
is generated by cmake internally, so a more-specific diagnostic is
desired.

Extend policy CMP0057 to cover this diagnostic.