Workaround SolarisStudio bug with libstdc++.
Using iostream and cxxabi.h together causes a compile failure

 .../CC-gcc/include/c++/4.8.2/cxxabi.h", line 131: Error: Only one of a set of overloaded functions can be extern "C".

https://community.oracle.com/thread/3644901

This only happens if the iostream include is before the cxxabi include.

The available solutions are: re-order the includes in
kwsys/SystemInformation.cxx, skip the test for KWSYS_CXX_HAS_CXXABI, or
cause it to fail by adding the include.

Do the latter for now.