Features: Exclude MSVC from link feature cycle rejection test
MSVC has no explicit flags to enable C++11/14/17 language standard
levels.  Instead each version of the compiler documents a list of
supported features that are always on.  Therefore no use of specific
features will adjust the CXX_STANDARD level required for any target.
Therefore we must exclude the RunCMake.CompileFeatures test case
"LinkImplementationFeatureCycle" because it can never fail.

Note that within the RunCMakeTest script no languages or toolchains are
enabled so we cannot test if(MSVC) directly.  Instead test it during
configuration of the test and pass in a variable telling RunCMakeTest
whether to activate this test case.