ninja: list outputs of CMake's generate step
List outputs of the generate step as outputs of the RERUN_CMAKE target.
Any outputs which are also inputs are not considered outputs because if
they are missing, Ninja will need to rerun CMake anyways since an input
file changed.

The "restat = 1" flag is also required in case outputs are not updated.

Fixes #15256.