Ninja: Use "deps = msvc" only for C and CXX
The "/showIncludes" flag is only available with MS C and C++ compilers,
and on compilers that "simulate" them (like Intel for Windows).  Fix our
logic to choose this type only for MS tools with these languages.  All
other cases need to use "deps = gcc" and define DEP_FILE in the build
rule.

Reported-by: Brian Smith <brian@briansmith.org>